(week-2 project)

+++ week-2/project/project_1.rs
@@ -15,13 +15,13 @@ fn main() {
 	So i researched on how to use the powers in Rust and i did a separate one just below using it.*/
 
 
-	let principal:f64 = 520_000_000.00;
+	let principal:f32 = 520_000_000.00;
 
-	let rate:f64 = 10.00;
+	let rate:f32 = 10.00;
 
-	let time:f64 = 5.00;
+	let time:f32 = 5.00;
 
-	let amount_2:f64 = principal * (1.00 + (rate / 100.00)).powf(time);
+	let amount_2:f32 = principal * (1.00 + (rate / 100.00)).powf(time);
 	let compound_interest = amount_2 - principal;
 
 	println!("The compound interest is {}", compound_interest);
